Refactor KM panel.
